+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## edge
+
+*   Depreceate $document.title and $document.title=.
+
 ## 0.0.8 2013-05-02
 
 *   Depreceate Document in favor of $document global.